Change data values

+++ b5.py
@@ -11,10 +11,10 @@ Xuất bạn tên có số calo tích lũy trong ngày là : số calo
 
 ngay = input("Ngay:")
 ten = input("Ten:")
-sang = int(input("So calo cho buoi sang:"))
-trua = int(input("So calo cho buoi trua:"))
-toi = int(input("So calo cho buoi toi:"))
-caloBurntToday = int(input("So calo da van dong trong ngay:"))
+sang = float(input("So calo cho buoi sang:"))
+trua = float(input("So calo cho buoi trua:"))
+toi = float(input("So calo cho buoi toi:"))
+caloBurntToday = float(input("So calo da van dong trong ngay:"))
 total = sang + trua + toi - caloBurntToday
 a = "{} co so calo tich luy trong ngay {} la {}"
 print(a.format(ten, ngay, total))
